[Improve][Connector-v2] Refactor the schema change code
